feat(navbar): add credits link
